+++ include/wx/msw/ole/dataobj.h
@@ -65,7 +65,7 @@ public:
     // StdFormat enumerations or a user-defined format)
   virtual bool IsSupportedFormat(wxDataFormat format) const = 0;
     // get the (total) size of data
-  virtual uint GetDataSize() const = 0;
+  virtual size_t GetDataSize() const = 0;
     // copy raw data to provided pointer
   virtual void GetDataHere(void *pBuf) const = 0;
 
@@ -93,7 +93,7 @@ public:
     { return (wxDataFormat) wxDataObject::Text; }
   virtual bool IsSupportedFormat(wxDataFormat format) const
     { return format == wxDataObject::Text || format == wxDataObject::Locale; }
-  virtual uint GetDataSize() const
+  virtual size_t GetDataSize() const
     { return m_strText.Len() + 1; } // +1 for trailing '\0'of course
   virtual void GetDataHere(void *pBuf) const
     { memcpy(pBuf, m_strText.c_str(), GetDataSize()); }
